Create the base of the UI.
Created PostsScreen that uses a LazyColumn of Cards to display the content.
Created temporary implementation of the and PostsViewModel that for now just fetches the data from the temporary implementation of PostsRepository.
DI: changed RepositoriesAndDataSourcesModule to an abstract class which is required to use @Binds annotation on abstract functions that bind an implementation of an interface.