Created Dto and Entity models of the User (and its internal objects) and Post.
Created mappers to convert Dto objects to Entity objects which will be required when storing the api data in the database.

Made slight corrections to the User domain model by changing Geo's lat and lng values to be nullable because api returns string values and they could possibly fail to convert to Doubles.
Also changed User's id to be of type Long instead of Int to remain consistent across the data and domain layer and avoid extra conversions.