Implemented exception handling in PostsRemoteDataSource and UsersRemoteDataSource.
Created ApiException sealed class two hold types of ApiExceptions with the purpose of returning clear and meaningful exception when they occur.

Created a ConnectivityChecker to be able to on the spot check whether device is online.

Updated a bit the PostsRepositoryImpl to implement retrieving User data from the api based on retrieved posts. TODO: save that data in the local database.