Implemented local database functionality:
Created PostsAppDatabase and Daos for both Posts and Users.
Created PostsLocalDataSource and UsersLocalDataSource classes to respectively handle saving and fetching of posts and users data.

Fixed a mistake in PostEntity where the primary key was not previously annotated with @PrimaryKey.

Created an additional mapper to handle mapping of a list of PostDto objects to a list of PostEntity objects.

Changed the PostsRepositoryImpl getPosts method to handle saving of retrieved posts and users data in the local database. Made an assumption that a refresh of data should essentially clean and download all new data.

Made fixes in UserEntity.kt: change the primary keys of CompanyEntity to be name and userId because Room does not allow autogenerating one of compose primary keys.
The same fix for AddressEntity. Removed addressEntity and companyEntity parameters from UserEntity. Fixed the same in UserMappers.kt.

Fixed an error from UsersRemoteDataSource.kt which was left there when testing.