Implemented fetching of Flow<List<Post>> from the PostsRepositoryImpl.
Implemented BasicResult.kt class to be able to return a simple Success or Failure result without needing to pass a long any data in case of success.

Changed Post domain model to have a nullable userName. This is required so it could better indicate not having found a user with such a name in the database instead of having an empty string.

Added PostEntity to Post mapper in PostMappers.kt

Changed the previous getPosts methods in PostsRepository to refreshPosts that returns a BasicResult, and created a new getPosts() method that now returns a flow of List of Post.

Fixed a naming mistake in UsersLocalDataSource.kt which had a method 'getPosts' that should have read 'getUsers'

Made small adjustments in PostsScreen.kt to handle the now nullable userName and added string resources to hold those placeholders values.

Adjusted PostsViewModel to now receive a BasicResult, but left the consequential actions as TODOs for the next change.