Implemented getting the flow of posts in the ui layer. Created a PostsUiState class to hold the state of the entire screen and an internal RefreshState that holds the state of refreshing of data.
Created a refreshData method in PostsViewModel which initializes a data refresh from the api.
Created uiState StateFlow that holds the state of the screen and is combination of flow of Posts and RefreshState.
Added showErrorDialog and showErrorDialogState state flows which is needed to be able to determine whether to show the dialog not only on the ui state but also on whether the user has dismissed the dialog or not.

Added required error strings to strings.xml